core: configure preetier for auto formatting - use preetier & prettier-plugin-solidity as dev dependency - preetify existing solidity files - re-compile & migrate existing contracts

+++ contracts/CourseMarketplace.sol
@@ -9,17 +9,15 @@ contract CourseMarketplace {
   }
 
   struct Course {
-    uint id; // default to uint256 - 32 bytes (slot 1)
-    uint price; // 32 bytes (slot 2)
-
+    uint256 id; // default to uint256 - 32 bytes (slot 1)
+    uint256 price; // 32 bytes (slot 2)
     /**
-     * 8 * 32 = 256 bits hexadecimal characters, keccak encoded that is unique to the user, purchaser of the course
+     * 8 * 32 = 256 bits hexadecimal characters,
+     * keccak256 encoded that is unique to the user, purchaser of the course
      */
     bytes32 proof; // 32 bytes (slot 3)
-    
     // user who has purchased the course
     address owner; // 20 bytes (slot 4)
-
     State state; // 1 byte (slot 4)
   }
 }
